Moved terminal access settings to security page
